refactor: internal api_client is now can be used importing /pkg/api
