+++ pkg/api/core_version.go
@@ -1,13 +1,15 @@
-package apiclient
+package api
 
 import "net/http"
 
-func (c HTTPClient) CoreGetVersion() (res struct {
+// CoreGetVersion API version details,
+// including some parts of the global datacenter config.
+func (c API) CoreGetVersion() (res struct {
 	Release string `json:"release"`
 	RepoID  string `json:"repoid"`
 	Version string `json:"version"`
 }, err error) {
-	err = c.sendPVERequest(PVERequest{
+	err = c.SendPVERequest(PVERequest{
 		Path:   "/api2/json/version",
 		Method: http.MethodGet,
 		Result: &res,
